# apps/server/cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for AgroMart backend services.",
        "title": "AgroMart API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/analytics/inventory": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Returns a placeholder inventory snapshot series",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Inventory snapshot",
                "responses": {
                    "200": {
                        "description": "snapshot series",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "invalid tenant/auth",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/analytics/kpis": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Returns KPI metrics for dashboard",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get KPI bundle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "From date (YYYY-MM-DD or RFC3339)",
                        "name": "from_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "To date (YYYY-MM-DD or RFC3339)",
                        "name": "to_date",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "description": "Low stock threshold",
                        "name": "threshold",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "description": "Top N selling products",
                        "name": "top",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "description": "Forward window days for inventory projection",
                        "name": "window_days",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "KPIs payload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "invalid tenant/auth",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/analytics/purchases": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Returns purchases cost time series",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Purchases cost series",
                "parameters": [
                    {
                        "type": "string",
                        "description": "From date (YYYY-MM-DD or RFC3339)",
                        "name": "from_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "To date (YYYY-MM-DD or RFC3339)",
                        "name": "to_date",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "day",
                            "month"
                        ],
                        "type": "string",
                        "description": "Group by: day or month",
                        "name": "group",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "time series points",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "invalid tenant/auth",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/analytics/sales": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Returns sales revenue time series",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Sales revenue series",
                "parameters": [
                    {
                        "type": "string",
                        "description": "From date (YYYY-MM-DD or RFC3339)",
                        "name": "from_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "To date (YYYY-MM-DD or RFC3339)",
                        "name": "to_date",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "day",
                            "month"
                        ],
                        "type": "string",
                        "description": "Group by: day or month",
                        "name": "group",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "time series points",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "invalid tenant/auth",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
}